feat: added initial structure for multiple result services - OResults, OFeed to be implemented
