+++ app/src/main/java/kolskypavel/ardfmanager/backend/room/enums/ResultServiceStatus.kt
@@ -2,8 +2,9 @@ package kolskypavel.ardfmanager.backend.room.enums
 
 enum class ResultServiceStatus(val value: Int) {
     DISABLED(0),
-    RUNNING(1),
-    NO_NETWORK(2),
-    UNAUTHORIZED(3),
-    ERROR(4);
+    INIT(1),
+    RUNNING(2),
+    NO_NETWORK(3),
+    UNAUTHORIZED(4),
+    ERROR(5);
 }
