+++ app/src/main/res/values/strings.xml
@@ -434,6 +434,7 @@
     <string name="result_service_type_robis_test" translatable="false">ROBis test</string>
 
     <string name="result_service_status_disabled">Disabled</string>
+    <string name="result_service_status_init">Init</string>
     <string name="result_service_status_running">Running</string>
     <string name="result_service_status_no_network">No network</string>
     <string name="result_service_status_unauthorized">Unauthorized</string>
@@ -559,6 +560,7 @@
 
     <string-array name="result_service_status">
         <item>@string/result_service_status_disabled</item>
+        <item>@string/result_service_status_init</item>
         <item>@string/result_service_status_running</item>
         <item>@string/result_service_status_no_network</item>
         <item>@string/result_service_status_unauthorized</item>
